Bug with PHP date format when convertFormat=true and widget format is not set
Problem: when `DatePicker` is configured with `convertFormat=true`, `format` not specified and `Yii::$app->formatter` specified the date in PHP format (for ex. 'php:Y-m-d'), the DatePicker incorrectly calls convertDateIcuToPhp() on that PHP format.

This happens due to the condition `if (isset($this->dateFormat) ... )` being always false, because the `$dateFormat` field does not exist (and never set) in both `InputWidget` and `DatePicker` classes. The local variable `$format` should have been used in the `if` instead.